move 'export_stuff' out of 'core_processes' (I feel like 'core_processes' means 'stuff for turning light into skeletons')
